chore: fix Data Pipelines TypeScript types exports (#463)
