+++ src/index.ts
@@ -1,7 +1,7 @@
 /**
  * This file serves as the entry point for the module customerio-reactnative.
  */
-export * from './cio-config';
 export * from './customerio-cdp';
 export * from './customerio-inapp';
 export * from './customerio-push';
+export * from './types';
