+++ src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './data-pipelines';
 export * from './InAppMessage';
 export * from './PushClickBehaviorAndroid';
 export * from './PushPermissionOptions';
